Version 0.11 * Added the Account option. This allows the user to save his data to the cloud. * When grouping by faction; each faction has a button for showing all that factions upgrade cards * Added some small help texts * User can now save filter/list options as Setups. Allowing for reuse of favourite combinations to be easily reused.

+++ src/components/upgrade.ts
@@ -15,7 +15,9 @@ export class Upgrade {
   constructor(private crewBuilderService: CrewBuilderService, private suit: SuitValueConverter) {}
 
   bind() {
-    this.added = this.model.upgrade && this.model.upgrade.name === this.upgrade.name;
+    if (this.model) {
+      this.added = this.model.upgrade && this.model.upgrade.name === this.upgrade.name;
+    }
     this.limitations = this.getLimitations();
   }
 
@@ -32,13 +34,17 @@ export class Upgrade {
   }
 
   addToModel() {
-    this.added= true;
-    this.crewBuilderService.addCrewModelUpgrade(this.model, this.upgrade);
+    if (this.model) {
+      this.added = true;
+      this.crewBuilderService.addCrewModelUpgrade(this.model, this.upgrade);
+    }
   }
 
   removeFromModel() {
-    this.added = false;
-    this.crewBuilderService.removeCrewModelUpgrade(this.model);
+    if (this.model) {
+      this.added = false;
+      this.crewBuilderService.removeCrewModelUpgrade(this.model);
+    }
   }
 
   getText(text):string {
